+++ index.html
@@ -16,7 +16,6 @@
     <div id='sidebar' class='fl fill-white'></div>
     <div id='main' class='unlimiter contain'>
         <div id='settings' class='animate modal modal-content'></div>
-        <div id='map' class='fill-black'></div>
     </div>
     <script src='js/bundle.js'></script>
 </body>
