+++ templates/map.html
@@ -1,0 +1,1 @@
+<div id='map' class='fill-black'></div>
